requirements.txt
Django>=3.1,<5
djangorestframework>=3.12
django-filter>=23.0
psycopg2-binary>=2.9
